requirements.txt
streamlit==1.57.0
tornado>=6.2
opencv-python-headless==4.13.0.92
numpy==2.4.4
pillow==12.2.0
insightface==0.7.3
onnxruntime==1.25.1
